requirements.txt
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.72
botocore==1.35.72
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==44.0.0
defusedxml==0.7.1
Django==4.2.10
django-allauth==0.61.1
django-anymail==10.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.4.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-storages==1.14.2
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.1
drf-spectacular==0.27.1
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
psycopg==3.1.18
psycopg-c==3.1.18
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
s3transfer==0.10.4
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
text-unidecode==1.3
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.27.1
watchfiles==1.0.0
websockets==14.1
wheel==0.45.1
whitenoise==6.6.0
